Implement JWTs
Successful implmenetation of JWTs use however receiving an error when logging via the form or passing valid login credentials via Postman.

The bcrypt.compare function is returning false even when the passwords match.

Datatypes have been checked and both values are passing as strings.

Need to resolve this issue in order for the secure login functionality to work properly.